+++ modules/math/src/main/java/com/opengamma/strata/math/impl/linearalgebra/LUDecompositionCommons.java
@@ -11,7 +11,7 @@ import org.apache.commons.math3.linear.RealMatrix;
 import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.collect.array.DoubleMatrix;
 import com.opengamma.strata.math.impl.util.CommonsMathWrapper;
-import com.opengamma.strata.math.rootfind.Decomposition;
+import com.opengamma.strata.math.linearalgebra.Decomposition;
 
 /**
  * This class is a wrapper for the
