+++ modules/math/src/main/java/com/opengamma/strata/math/impl/linearalgebra/SVDecompositionResult.java
@@ -6,7 +6,7 @@
 package com.opengamma.strata.math.impl.linearalgebra;
 
 import com.opengamma.strata.collect.array.DoubleMatrix;
-import com.opengamma.strata.math.rootfind.DecompositionResult;
+import com.opengamma.strata.math.linearalgebra.DecompositionResult;
 
 /**
  * Contains the results of SV matrix decomposition.
