+++ modules/math/src/main/java/com/opengamma/strata/math/impl/rootfinding/newton/NewtonDefaultVectorRootFinder.java
@@ -6,7 +6,7 @@
 package com.opengamma.strata.math.impl.rootfinding.newton;
 
 import com.opengamma.strata.math.impl.linearalgebra.LUDecompositionCommons;
-import com.opengamma.strata.math.rootfind.Decomposition;
+import com.opengamma.strata.math.linearalgebra.Decomposition;
 
 /**
  * A root finder that attempts find the multi-dimensional root of a series of N equations with N variables (a square problem).
