+++ modules/math/src/main/java/com/opengamma/strata/math/linearalgebra/Decomposition.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.math.rootfind;
+package com.opengamma.strata.math.linearalgebra;
 
 import java.util.function.Function;
