+++ modules/math/src/main/java/com/opengamma/strata/math/linearalgebra/DecompositionResult.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.math.rootfind;
+package com.opengamma.strata.math.linearalgebra;
 
 import com.opengamma.strata.collect.array.DoubleArray;
 import com.opengamma.strata.collect.array.DoubleMatrix;
