+++ modules/math/src/main/java/com/opengamma/strata/math/linearalgebra/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright (C) 2018 - present by OpenGamma Inc. and the OpenGamma group of companies
+ *
+ * Please see distribution for license.
+ */
+
+/**
+ * Linear algebra.
+ */
+package com.opengamma.strata.math.linearalgebra;
