+++ modules/math/src/main/java/com/opengamma/strata/math/rootfind/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright (C) 2018 - present by OpenGamma Inc. and the OpenGamma group of companies
+ *
+ * Please see distribution for license.
+ */
+
+/**
+ * Root finding.
+ */
+package com.opengamma.strata.math.rootfind;
