+++ modules/math/src/test/java/com/opengamma/strata/math/impl/linearalgebra/SVDecompositionCommonsTest.java
@@ -9,7 +9,7 @@ import org.testng.annotations.Test;
 
 import com.opengamma.strata.math.impl.matrix.CommonsMatrixAlgebra;
 import com.opengamma.strata.math.impl.matrix.MatrixAlgebra;
-import com.opengamma.strata.math.rootfind.Decomposition;
+import com.opengamma.strata.math.linearalgebra.Decomposition;
 
 /**
  * Test.
